testCloudCheckpointForceWithNumToKeep
